% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{clean_mvt}
\alias{clean_mvt}
\title{Stop all running vector tile servers}
\usage{
clean_mvt()
}
\description{
As you use serve_mvt or view_mvt servers will accumulate in child processes.
This function kills all child processes serving tiles.
}
