% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtview-package.R
\docType{package}
\name{mvtview-package}
\alias{mvtview}
\alias{mvtview-package}
\title{mvtview: View and Serve Mapbox Vector Tile Databases}
\description{
View and Serve Mapbox Vector Tile Databases for mapping development tasks.
}
\author{
\strong{Maintainer}: Miles McBain \email{miles.mcbain@gmail.com} (\href{https://orcid.org/0000-0003-2865-2548}{ORCID})

}
\keyword{internal}
