% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serve_mvt}
\alias{serve_mvt}
\title{Serve a .mbtiles database of vectortiles}
\usage{
serve_mvt(tiles_path, .serve_mode = "in-memory")
}
\arguments{
\item{tiles_path}{The path to an .mbtiles file.}

\item{.serve_mode}{The way in which the server handles the vector tiles
database. "in-memory" is the default and it will read the entire tile database
into R as a tibble. "disk" will read tiles from the .mbtiles file as an
SQLite database  from disk. The default is more performant. Use "disk" only if
you have a large vector tileset that would consume too much memory to hold in
RAM at once.}
}
\description{
Starts a web server in a background R session serving vector tiles from a
supplied .mbtiles file.
}
\seealso{
start_mvt_server for more control of server behaviour.
}
