% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{view_mvt}
\alias{view_mvt}
\title{View a local vector tileset on a map}
\usage{
view_mvt(
  tiles_path,
  get_fill_color = "#FFFFFF70",
  get_line_color = "#ffffffff",
  get_line_width = 2,
  line_width_units = "pixels",
  get_point_radius = 2,
  point_radius_units = "pixels",
  stroked = TRUE,
  tooltip = TRUE,
  pickable = TRUE,
  ...,
  .serve_mode = "in-memory"
)
}
\arguments{
\item{tiles_path}{The path to an .mbtiles file.}

\item{get_fill_color}{the fill colour of plotted features.}

\item{get_line_color}{the line colour of plotted features.}

\item{get_line_width}{the line width of plotted features (in pixels by default).}

\item{line_width_units}{the units of the value supplied in \code{get_line_width}.
"meters" may be preferred in some cases.}

\item{get_point_radius}{the radius of plotted point features (in pixels by default).}

\item{point_radius_units}{the units of the value supplied in \code{get_point_radius}.
"meters" may be preferred in some cases.}

\item{stroked}{use a line on the borders of polygons or points? TRUE by default.}

\item{tooltip}{generate a tooltip for feature attributes? TRUE by default.}

\item{pickable}{allow map to react to features that get mouse hover? Needs to be
enabled to view tooltips. TRUE by default.}

\item{...}{further arguments forwarded to \code{\link[rdeck:mvt_layer]{rdeck::add_mvt_layer()}}.}

\item{.serve_mode}{The way in which the server handles the vector tiles
database. "in-memory" is the default and it will read the entire tile database
into R as a tibble. "disk" will read tiles from the .mbtiles file as an
SQLite database  from disk. The default is more performant. Use "disk" only if
you have a large vector tileset that would consume too much memory to hold in
RAM at once.}
}
\description{
Given a local .mtiles file containing a vector tiles database, this function
will start a local development server to serve the tiles and then return a htmlwidget map
that displays the tileset.
}
\details{
The map is powered by the awesome {rdeck} package, which is highly
recommended for making interactive WebGL maps in R.

The graphics options of this function are passed directed to '
\code{\link[rdeck:mvt_layer]{rdeck::add_mvt_layer()}}, and so suppourt {rdeck} color scales based on
attributes. See the {rdeck} helpfile for more detailed ' descriptions.

The graphics parameters apply only to relvant geometries. For ' example:
'fill color' is not used for line string features.
}
